# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScholarAI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": [
          "Profile and privacy"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile and privacy"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "put": {
        "tags": [
          "Profile and privacy"
        ],
        "summary": "Settings Update",
        "operationId": "settings_update_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scholarships/manual": {
      "post": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Manual",
        "operationId": "manual_api_scholarships_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScholarshipInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scholarships/from-url": {
      "post": {
        "tags": [
          "Scholarships"
        ],
        "summary": "From Url",
        "operationId": "from_url_api_scholarships_from_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UrlInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scholarships/from-document": {
      "post": {
        "tags": [
          "Scholarships"
        ],
        "summary": "From Document",
        "operationId": "from_document_api_scholarships_from_document_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scholarships/{scholarship_id}": {
      "get": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Detail",
        "operationId": "detail_api_scholarships__scholarship_id__get",
        "parameters": [
          {
            "name": "scholarship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scholarship Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Update",
        "operationId": "update_api_scholarships__scholarship_id__put",
        "parameters": [
          {
            "name": "scholarship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scholarship Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScholarshipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scholarships/{scholarship_id}/retry": {
      "post": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Retry",
        "operationId": "retry_api_scholarships__scholarship_id__retry_post",
        "parameters": [
          {
            "name": "scholarship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scholarship Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/requirements/{requirement_id}": {
      "patch": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Admin Update",
        "operationId": "admin_update_api_admin_requirements__requirement_id__patch",
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Criterion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/scholarships/{scholarship_id}": {
      "patch": {
        "tags": [
          "Scholarships"
        ],
        "summary": "Admin Deactivate",
        "operationId": "admin_deactivate_api_admin_scholarships__scholarship_id__patch",
        "parameters": [
          {
            "name": "scholarship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scholarship Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScholarshipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Criterion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "category": {
            "type": "string",
            "enum": [
              "academic",
              "language",
              "degree",
              "field_of_study",
              "age",
              "nationality",
              "graduation",
              "experience",
              "research",
              "technical",
              "documents",
              "other"
            ],
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "criterion_type": {
            "type": "string",
            "enum": [
              "numeric_min",
              "numeric_max",
              "boolean",
              "exact_match",
              "one_of",
              "date_before",
              "date_after",
              "document_required",
              "semantic_match",
              "manual_review",
              "any_of",
              "all_of"
            ],
            "title": "Criterion Type"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "default": ""
          },
          "operator": {
            "type": "string",
            "title": "Operator",
            "default": ""
          },
          "required_value": {
            "title": "Required Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "mandatory": {
            "type": "boolean",
            "title": "Mandatory",
            "default": true
          },
          "weight": {
            "type": "number",
            "minimum": 0.0,
            "title": "Weight",
            "default": 1
          },
          "source_text": {
            "type": "string",
            "minLength": 1,
            "title": "Source Text"
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.9
          },
          "admin_verified": {
            "type": "boolean",
            "title": "Admin Verified",
            "default": false
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/Criterion"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "title",
          "criterion_type",
          "source_text"
        ],
        "title": "Criterion"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProfileInput": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name",
            "default": ""
          },
          "nationality": {
            "type": "string",
            "title": "Nationality",
            "default": ""
          },
          "country_of_residence": {
            "type": "string",
            "title": "Country Of Residence",
            "default": ""
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "university": {
            "type": "string",
            "title": "University",
            "default": ""
          },
          "degree": {
            "type": "string",
            "title": "Degree",
            "default": ""
          },
          "field_of_study": {
            "type": "string",
            "title": "Field Of Study",
            "default": ""
          },
          "gpa": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpa"
          },
          "gpa_scale": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Gpa Scale",
            "default": 4
          },
          "graduation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Date"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2200.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          },
          "english_test": {
            "type": "string",
            "title": "English Test",
            "default": ""
          },
          "english_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "English Score"
          },
          "test_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Date"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Skills"
          },
          "research": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Research"
          },
          "experience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Experience"
          },
          "courses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Courses"
          },
          "target_degree": {
            "type": "string",
            "title": "Target Degree",
            "default": "Master's"
          },
          "target_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Fields"
          },
          "preferred_countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Countries"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileInput"
      },
...
      "ScholarshipInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 250,
            "minLength": 3,
            "title": "Name"
          },
          "text": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 10,
            "title": "Text"
          },
          "university": {
            "type": "string",
            "title": "University",
            "default": ""
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": ""
          },
          "degree_level": {
            "type": "string",
            "title": "Degree Level",
            "default": "Master's"
          },
          "field": {
            "type": "string",
            "title": "Field",
            "default": ""
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "funding_type": {
            "type": "string",
            "title": "Funding Type",
            "default": "Not specified"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "text"
        ],
        "title": "ScholarshipInput"
      },
      "ScholarshipUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 250,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ScholarshipUpdate"
      },
      "SettingsInput": {
        "properties": {
          "ai_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Enabled"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "current_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Password"
          },
          "new_password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 12
              },
              {
                "type": "null"
              }
            ],
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsInput"
      },
...
      "UrlInput": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2000,
            "title": "Url"
          },
          "name": {
            "type": "string",
            "maxLength": 250,
            "title": "Name",
            "default": "Imported scholarship"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "UrlInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}